Fix tmdb NPE

+++ Europium/Services/Apis/QBitTorrent/TorrentInfo.cs
@@ -10,7 +10,7 @@ public class TorrentInfo
 	public string? State { get; set; }
 	public long Size { get; set; }
 	public float Progress { get; set; }
-	public int Eta { get; set; }
+	public long Eta { get; set; }
 	[JsonProperty("added_on")]
 	public long AddedOn { get; set; }
 	[JsonProperty("completion_on")]
